requirements.txt
boto3
minio
flask
flask-socketio
eventlet
numpy
opencv-python-headless
PyYAML
requests
tqdm
